# swagger.yml
swagger: '2.0'
info:
  description: DevTest
  version: 1.0.0
  title: DevTest
  # put the contact info for your development or API team
  contact:
    email: javigaton@gmail.com

  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html

# tags are used for organizing operations
...
paths:
...
  /users/{username}/teams:
    get:
      security:
        - BearerCookie: []
      tags:
      - user
      summary: Returns all teams of a user.
      operationId: GetTeamsOfUser
      description: Returns all teams of a user.
      produces:
      - application/json
      parameters:
      - in: path
        name: username
        description: Username of the user to get their teams
        required: true
        type: string
      responses:
        200:
          description: teams found
          schema:
            type: array
            items:
              $ref: '#/definitions/Team'
        400:
          $ref: '#/responses/BadRequestError'
        403:
          $ref: '#/responses/ForbiddenError'
        500:
          $ref: '#/responses/InternalServerError'
    post:
      security:
        - BearerCookie: []
      tags:
      - team
      - user
      summary: adds a team
      operationId: PostTeam
      description: Adds a team to the system
      consumes:
      - application/json
      produces:
      - application/json
      parameters:
      - in: path
        name: username
        description: Username of the user to get their teams
        required: true
        type: string
      - in: body
        name: team
        description: Team item to add
        required: true
        schema:
          $ref: '#/definitions/Team'
      responses:
        201:
          description: team created
          schema:
            $ref: '#/definitions/Team'
        400:
          $ref: '#/responses/BadRequestError'
        403:
          $ref: '#/responses/ForbiddenError'
        409:
          $ref: '#/responses/ConflictError'
        500:
          $ref: '#/responses/InternalServerError'
...
  /teams:
    get:
      security:
        - BearerCookie: []
      tags:
      - team
      summary: Returns all teams.
      operationId: GetTeams
      description: Returns all teams.
      produces:
      - application/json
      responses:
        200:
          description: teams found
          schema:
            type: array
            items:
              $ref: '#/definitions/Team'
        400:
          $ref: '#/responses/BadRequestError'
        403:
          $ref: '#/responses/ForbiddenError'
        500:
          $ref: '#/responses/InternalServerError'
  /teams/{teamname}:
    get:
      security:
        - BearerCookie: []
      tags:
      - team
      summary: Finds a team by its teamname
      operationId: getTeam
      description: Finds a team by its teamname
      produces:
      - application/json
      parameters:
      - in: path
        name: teamname
        description: Teamname of the team to find
        required: true
        type: string
      responses:
        200:
          description: team found
          schema:
            $ref: '#/definitions/Team'
        400:
          $ref: '#/responses/BadRequestError'
        403:
          $ref: '#/responses/ForbiddenError'
        410:
          $ref: '#/responses/GoneError'
        500:
          $ref: '#/responses/InternalServerError'
    put:
      security:
        - BearerCookie: []
      tags:
      - team
      summary: Modifies the team <teamname>
      operationId: putTeam
      description: Modifies the team <teamname>
      consumes:
      - application/json
      parameters:
      - in: path
        name: teamname
        description: Teamname of the team to modify its information
        required: true
        type: string
      - in: body
        name: team
        required: true
        description: Team information updated
        schema:
          $ref: '#/definitions/Team'
      responses:
        200:
          description: team updated
        400:
          $ref: '#/responses/BadRequestError'
        403:
          $ref: '#/responses/ForbiddenError'
        409:
          $ref: '#/responses/ConflictError'
        410:
          $ref: '#/responses/GoneError'
        500:
          $ref: '#/responses/InternalServerError'
    delete:
      security:
        - BearerCookie: []
      tags:
      - team
      summary: Deletes a team by its teamname
      operationId: deleteTeam
      description: Deletes a team by its teamname
      parameters:
      - in: path
        name: teamname
        description: Teamname of the team to delete
        required: true
        type: string
      responses:
        200:
          description: team found and deleted
        400:
          $ref: '#/responses/BadRequestError'
        403:
          $ref: '#/responses/ForbiddenError'
        410:
          $ref: '#/responses/GoneError'
        500:
          $ref: '#/responses/InternalServerError'
...
  /teams/{teamname}/users:
    get:
      security:
        - BearerCookie: []
      tags:
      - team
      summary: Returns all users that are in a team
      operationId: GetUsersFromTeam
      description: Returns all users that are in a team
      produces:
      - application/json
      parameters:
      - in: path
        name: teamname
        description: Teamname of the team to get its users
        required: true
        type: string
      responses:
        200:
          description: users found
          schema:
            type: array
            items:
              $ref: '#/definitions/User'
        400:
          $ref: '#/responses/BadRequestError'
        403:
          $ref: '#/responses/ForbiddenError'
        410:
          $ref: '#/responses/GoneError'
        500:
          $ref: '#/responses/InternalServerError'   
  /teams/{teamname}/users/{username}:
    get:
      security:
        - BearerCookie: []
      tags:
      - team
      summary: Returns a user that is in a team
      operationId: GetUserFromTeam
      description: Returns a user that is in a team
      produces:
      - application/json
      parameters:
      - in: path
        name: teamname
        description: Teamname of the team to get its user
        required: true
        type: string
      - in: path
        name: username
        description: Username of the user to check if it is in the team
        required: true
        type: string
      responses:
        200:
          description: user found
          schema:
            $ref: '#/definitions/User'
        400:
          $ref: '#/responses/BadRequestError'
        403:
          $ref: '#/responses/ForbiddenError'
        410:
          $ref: '#/responses/GoneError'
        500:
          $ref: '#/responses/InternalServerError'
    delete:
      security:
        - BearerCookie: []
      tags:
      - team
      summary: Deletes user {username} from team {teamname}
      operationId: DeleteUserFromTeam
      description: Deletes user {username} from team {teamname}
      produces:
      - application/json
      parameters:
      - in: path
        name: teamname
        description: Teamname of the team to modify
        required: true
        type: string
      - in: path
        name: username
        description: Username of the user to delete from team
        required: true
        type: string
      responses:
        200:
          description: User deleted from team
        400:
          $ref: '#/responses/BadRequestError'
        403:
          $ref: '#/responses/ForbiddenError'
        410:
          $ref: '#/responses/GoneError'
        500:
          $ref: '#/responses/InternalServerError'
  /teams/{teamname}/questions:
    get:
      security:
        - BearerCookie: []
      tags:
      - team
      summary: Returns all questions that the team administers
      operationId: GetQuestionsFromTeam
      description: Returns all questions that the team administers
      produces:
      - application/json
      parameters:
      - in: path
        name: teamname
        description: Teamname of the team to get its questions
        required: true
        type: string
      - in: query
        name: tags
        type: array
        collectionFormat: pipes
        items:
          type: array
          collectionFormat: csv
          items:
            type: string
      - in: query
        name: likeTitle
        type: string
      - in: query
        name: orderby
        type: string
        enum: [newDate, oldDate, moreFav, lessFav, moreTime, lessTime]
        description: Indicates which element is first returned. In case of tie it unties with newdate first
      - in: query
        name: limit
        type: integer
        description: max number of elements to be returned
        minimum: 0
      - in: query
        name: offset
        type: integer
        description: first elements to be skipped at being returned
        minimum: 0
      responses:
        200:
          description: questions found
          schema:
            type: array
            items:
              $ref: '#/definitions/Question'
        400:
          $ref: '#/responses/BadRequestError'
        403:
          $ref: '#/responses/ForbiddenError'
        410:
          $ref: '#/responses/GoneError'
        500:
          $ref: '#/responses/InternalServerError'
  /teams/{teamname}/questions/{questionid}:
    get:
      security:
        - BearerCookie: []
      tags:
      - team
      summary: Returns a question that the team administers
      operationId: GetQuestionFromTeam
      description: Returns a question that the team administers
      produces:
      - application/json
      parameters:
      - in: path
        name: teamname
        description: Teamname of the team to get its question
        required: true
        type: string
      - in: path
        name: questionid
        description: Id of the question to find
        required: true
        type: integer
      responses:
        200:
          description: question found
          schema:
            $ref: '#/definitions/Question'
        400:
          $ref: '#/responses/BadRequestError'
        403:
          $ref: '#/responses/ForbiddenError'
        410:
          $ref: '#/responses/GoneError'
        500:
          $ref: '#/responses/InternalServerError'
  /teams/{teamname}/tests:
    get:
      security:
        - BearerCookie: []
      tags:
      - team
      summary: Returns all tests that the team administers
      operationId: GetTestsFromTeam
      description: Returns all tests that the team administers
      produces:
      - application/json
      parameters:
      - in: path
        name: teamname
        description: Teamname of the team to get its tests
        required: true
        type: string
      - in: query
        name: tags
        type: array
        collectionFormat: pipes
        items:
          type: array
          collectionFormat: csv
          items:
            type: string
      - in: query
        name: likeTitle
        type: string
      - in: query
        name: orderby
        type: string
        enum: [newDate, oldDate, moreFav, lessFav, moreTime, lessTime]
        description: Indicates which element is first returned. In case of tie it unties with newdate first
      - in: query
        name: limit
        type: integer
        description: max number of elements to be returned
        minimum: 0
      - in: query
        name: offset
        type: integer
        description: first elements to be skipped at being returned
        minimum: 0
      responses:
        200:
          description: tests found
          schema:
            type: array
            items:
              $ref: '#/definitions/Test'
        400:
          $ref: '#/responses/BadRequestError'
        403:
          $ref: '#/responses/ForbiddenError'
        410:
          $ref: '#/responses/GoneError'
        500:
          $ref: '#/responses/InternalServerError'
  /teams/{teamname}/tests/{testid}:
    get:
      security:
        - BearerCookie: []
      tags:
      - team
      summary: Returns a test that the team administers
      operationId: GetTestFromTeam
      description: Returns a test that the team administers
      produces:
      - application/json
      parameters:
      - in: path
        name: teamname
        description: Teamname of the team to get its question
        required: true
        type: string
      - in: path
        name: testid
        description: Id of the test to find
        required: true
        type: integer
      responses:
        200:
          description: test found
          schema:
            $ref: '#/definitions/Test'
        400:
          $ref: '#/responses/BadRequestError'
        403:
          $ref: '#/responses/ForbiddenError'
        410:
          $ref: '#/responses/GoneError'
        500:
          $ref: '#/responses/InternalServerError'
...
  /teams/{teamname}/invitedTests:
    get:
      security:
        - BearerCookie: []
      tags:
      - team
      summary: Returns all publishedTests that the team is invited to
      operationId: GetInvitedTestsFromTeam
      description: Returns all publishedTests that the team is invited to
      produces:
      - application/json
      parameters:
      - in: path
        name: teamname
        description: Teamname of the team to get its publishedTests is invited to
        required: true
        type: string
      - in: query
        name: tags
        type: array
        collectionFormat: pipes
        items:
          type: array
          collectionFormat: csv
          items:
            type: string
      - in: query
        name: likeTitle
        type: string
      - in: query
        name: orderby
        type: string
        enum: [newDate, oldDate, moreFav, lessFav, moreTime, lessTime]
        description: Indicates which element is first returned. In case of tie it unties with newdate first
      - in: query
        name: limit
        type: integer
        description: max number of elements to be returned
        minimum: 0
      - in: query
        name: offset
        type: integer
        description: first elements to be skipped at being returned
        minimum: 0
      responses:
        200:
          description: publishedTests found
          schema:
            type: array
            items:
              $ref: '#/definitions/Test'
        400:
          $ref: '#/responses/BadRequestError'
        403:
          $ref: '#/responses/ForbiddenError'
        410:
          $ref: '#/responses/GoneError'
        500:
          $ref: '#/responses/InternalServerError'
...
responses:
  BadRequestError:
    description: Incorrect Request, or invalida data
    schema:
      $ref: '#/definitions/Error'
  ForbiddenError:
    description: Not authorized to this content
    schema:
      $ref: '#/definitions/Error'
  ConflictError:
    description: A user with same username/email already exists
    schema:
      $ref: '#/definitions/Error'
  GoneError:
    description: That resource does not exist
    schema:
      $ref: '#/definitions/Error'
  InternalServerError:
    description: Internal error
definitions:
...
  Team:
    type: object
    required:
    - teamname
    - soloProfesores
    properties:
      teamname:
        type: string
        example: devtestTeam
        pattern: '^[^@ \t\r\n]+$'
      description:
        type: string
        example: DevTest Team
      soloProfesores:
        type: boolean
        example: true
...
  Error:
    type: object
    required:
    - message
    properties:
      message:
        type: string
        example: Object not found
...
securityDefinitions:
  BearerCookie:
    type: apiKey
    in: header # Realmente en cookie, pero las cookies estan en header
    name: Cookie # La cookie Bearer-Cookie
  ReAuthCookie:
    type: apiKey
    in: header # Realmente en cookie, pero las cookies estan en header
    name: Cookie # La cookie ReAuth-Cookie
  NoRegistered:
    type: apiKey
    in: header
    name: NotLoggedIn # Esta API KEY valdra true si el cliente quiere que se procese la peticion como si no estuviera registrado
host: localhost
basePath: /DevTest
schemes:
 - https